Split out gotify class to its own module
