+++ Dockerfile
@@ -12,7 +12,8 @@ ADD requirements.txt .
 RUN python -m pip install -r requirements.txt
 
 WORKDIR /app
-ADD alertify.py /app
+ADD src/alertify.py /app
+ADD src/gotify.py /app
 
 # Switching to a non-root user, please refer to https://aka.ms/vscode-docker-python-user-rights
 RUN useradd appuser && chown -R appuser /app
